Man kann nun die Persönlichen Daten für den Pannenservice ändern und dafür die Handynummer in einer long long gespeichert

+++ CarAssist/CarAssist/Profile.h
@@ -11,7 +11,7 @@
 
 @interface Profile : NSObject
 @property (nonatomic) NSString *vorname, *nachname,*emailAdresse,*lieblingsKEKSEEE,*ADAClicence, *homeTown;
-@property (nonatomic) NSInteger mobilenumber;
+@property (nonatomic) long long mobilenumber;
 @property (nonatomic) NSMutableArray *carList;
 @property (nonatomic) Car *car;
